FCFS works

+++ src/simulator/scheduler/PCB.java
@@ -77,7 +77,7 @@ public class PCB {
         return this.arrivalTime;
     }
 
-    protected void setArrivalTime (int arrivalTime) {
+    protected void setArrivalTime (float arrivalTime) {
         this.arrivalTime = arrivalTime;
     }
 }
